+++ example/android/app/src/main/kotlin/com/fintays/emoji_picker_flutter_example/MainActivity.kt
@@ -1,6 +1,0 @@
-package com.fintays.emoji_picker_flutter_example
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}