+++ lib/emoji_picker_flutter.dart
@@ -1,7 +1,8 @@
 library emoji_picker_flutter;
 
-export 'package:emoji_picker_flutter/src/emoji_picker.dart';
-export 'package:emoji_picker_flutter/src/emoji_picker_builder.dart';
 export 'package:emoji_picker_flutter/src/category_icons.dart';
 export 'package:emoji_picker_flutter/src/config.dart';
 export 'package:emoji_picker_flutter/src/emoji.dart';
+export 'package:emoji_picker_flutter/src/emoji_picker.dart';
+export 'package:emoji_picker_flutter/src/emoji_picker_builder.dart';
+export 'package:emoji_picker_flutter/src/emoji_picker_utils.dart';