+++ lib/src/emoji_skin_tones.dart
@@ -1,0 +1,22 @@
+/// Alternative skin tones of Emoji
+class SkinTone {
+  SkinTone._();
+
+  /// Light Skin Tone
+  static const String light = '🏻';
+
+  /// Medium-Light Skin Tone
+  static const String mediumLight = '🏼';
+
+  /// Medium Skin Tone
+  static const String medium = '🏽';
+
+  /// Medium-Dark Skin Tone
+  static const String mediumDark = '🏾';
+
+  /// Dark Skin Tone
+  static const String dark = '🏿';
+
+  /// Return all values as Array
+  static const values = [light, mediumLight, medium, mediumDark, dark];
+}