+++ macos/Classes/EmojiPickerFlutterPlugin.swift
@@ -1,0 +1,19 @@
+import Cocoa
+import FlutterMacOS
+
+public class EmojiPickerFlutterPlugin: NSObject, FlutterPlugin {
+  public static func register(with registrar: FlutterPluginRegistrar) {
+    let channel = FlutterMethodChannel(name: "emoji_picker_flutter", binaryMessenger: registrar.messenger)
+    let instance = EmojiPickerFlutterPlugin()
+    registrar.addMethodCallDelegate(instance, channel: channel)
+  }
+
+  public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
+    switch call.method {
+    case "getPlatformVersion":
+      result("macOS " + ProcessInfo.processInfo.operatingSystemVersionString)
+    default:
+      result(FlutterMethodNotImplemented)
+    }
+  }
+}