+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## 1.1.2
+- Fix a issue with macos platform 👨‍🔧
+- Close Skin-Tone dialog on several other timings
+
+## 1.1.1
+* Fix parsing of legacy Emoji Recent-List 😵
+
+## 1.1.0
+* New feature: EmojiPickerUtils provide access to recent emojis, search emoji and adding emoji to recently-used list
+* New feature: Skin-Tone Support
+* New feature: Emoji-List Versioning (force update users cached emoji's if necessary between versions)
+* Add Support for MacOS platform 🖥
+
+## 1.0.8
+
+* Fix to avoid unnecessary rebuilds
+* Re-add Activities Category 
+
 ## 1.0.7
 
 * Align center NoRecent-Text